Moved methods closer to type defintions.
